+++ src/backend/InvenTree/InvenTree/settings.py
@@ -194,9 +194,17 @@ PLUGINS_MANDATORY = get_setting(
     'INVENTREE_PLUGINS_MANDATORY', 'plugins_mandatory', typecast=list, default_value=[]
 )
 
+if PLUGINS_MANDATORY:
+    logger.info('Mandatory plugins: %s', PLUGINS_MANDATORY)
+
 PLUGINS_INSTALL_DISABLED = get_boolean_setting(
     'INVENTREE_PLUGIN_NOINSTALL', 'plugin_noinstall', False
 )
+
+if not PLUGINS_ENABLED:
+    PLUGINS_INSTALL_DISABLED = (
+        True  # If plugins are disabled, also disable installation
+    )
 
 PLUGIN_FILE = config.get_plugin_file()
 
@@ -831,10 +839,15 @@ GLOBAL_CACHE_ENABLED = is_global_cache_enabled()
 
 CACHES = {'default': get_cache_config(GLOBAL_CACHE_ENABLED)}
 
-_q_worker_timeout = int(
+BACKGROUND_WORKER_TIMEOUT = int(
     get_setting('INVENTREE_BACKGROUND_TIMEOUT', 'background.timeout', 90)
 )
 
+# Set the retry time for background workers to be slightly longer than the worker timeout, to ensure that workers have time to timeout before being retried
+BACKGROUND_WORKER_RETRY = max(
+    int(get_setting('INVENTREE_BACKGROUND_RETRY', 'background.retry', 300)),
+    BACKGROUND_WORKER_TIMEOUT + 120,
+)
 
 # Prevent running multiple background workers if global cache is disabled
 # This is to prevent scheduling conflicts due to the lack of a shared cache
@@ -848,16 +861,18 @@ BACKGROUND_WORKER_COUNT = (
 if 'sqlite' in DB_ENGINE:
     BACKGROUND_WORKER_COUNT = 1
 
+BACKGROUND_WORKER_ATTEMPTS = int(
+    get_setting('INVENTREE_BACKGROUND_MAX_ATTEMPTS', 'background.max_attempts', 5)
+)
+
 # django-q background worker configuration
 Q_CLUSTER = {
     'name': 'InvenTree',
     'label': 'Background Tasks',
     'workers': BACKGROUND_WORKER_COUNT,
-    'timeout': _q_worker_timeout,
-    'retry': max(120, _q_worker_timeout + 30),
-    'max_attempts': int(
-        get_setting('INVENTREE_BACKGROUND_MAX_ATTEMPTS', 'background.max_attempts', 5)
-    ),
+    'timeout': BACKGROUND_WORKER_TIMEOUT,
+    'retry': BACKGROUND_WORKER_RETRY,
+    'max_attempts': BACKGROUND_WORKER_ATTEMPTS,
     'save_limit': 1000,
     'queue_limit': 50,
     'catch_up': False,