+++ src/backend/InvenTree/company/api.py
@@ -184,6 +184,7 @@ class ManufacturerPartMixin(SerializerContextMixin):
 
 
 class ManufacturerPartList(
+    DataExportViewMixin,
     ManufacturerPartMixin,
     SerializerContextMixin,
     OutputOptionsMixin,